xen: define BIOVEC_PHYS_MERGEABLE()
When running in Xen domain with device access, we need to make sure
the block subsystem doesn't merge requests across pages which aren't
machine physically contiguous.  To do this, we define our own
BIOVEC_PHYS_MERGEABLE.  When CONFIG_XEN isn't enabled, or we're not
running in a Xen domain, this has identical behaviour to the normal
implementation.  When running under Xen, we also make sure the
underlying machine pages are the same or adjacent.

[ Impact: allow Xen control of bio merging ]

Signed-off-by: Jeremy Fitzhardinge <jeremy.fitzhardinge@citrix.com>