pv-ops: register xen pci notifier
Register the notifier to handle hot-plug devices and SR-IOV devices
for Xen hypervisor. When a device is hot added or removed, it adds
or removes it to Xen via hypercalls.

Changes in v3:
It isn't necessarily to explicitly initialize elements to 0, because
initializer will do it implicitly. Remove the unnecessary initilization.

Changes in v2:
Remove inline #ifdef and the awkward dangling else/#endif construction,
and rather than using memset, use variable declaration and initializer
to assign the elements in xen_add_device.

Signed-off-by: Weidong Han <weidong.han@intel.com>