xen/netback: use smart polling instead of event notification
Netback will not notify netfront until it finds that the netfront has
stopped polling.

Netback will set a flag in xenstore to indicate whether netback
supports the smart polling feature.  If there is only one side
supporting it, the communication mechanism will fall back to default,
and the new feature will not be used. The feature is enabled only
when both sides have the flag set in xenstore.

Signed-off-by: Dongxiao Xu <dongxiao.xu@intel.com>
Signed-off-by: Jeremy Fitzhardinge <jeremy.fitzhardinge@citrix.com>