xen/netfront: Use smart polling instead of event notification.
Netfront uses a timer to poll the ring instead of event channel
notification. If guest is transferring data, the timer will start working
and periodicaly send/receive data from ring. If guest is idle and no
data is transferring, the timer will stop working automatically. It will
restart again once there is new data transferring.

We set a feature flag in xenstore to indicate whether the netfront
support this feature. If there is only one side supporting it, the
communication mechanism will fall back to default, and the new feature
will not be used. The feature is enabled only when both sides have the
flag set in xenstore.

We set the default polling frequency is 1KHz, and implement get_coalesce()
and set_coalesce() interface for users to get and set the polling
frequency.

Signed-off-by: Dongxiao Xu <dongxiao.xu@intel.com>
Signed-off-by: Jeremy Fitzhardinge <jeremy.fitzhardinge@citrix.com>