drm: make sure all pages in vmalloc area are really DMA-ready
Under Xen, vmalloc_32() isn't guaranteed to return pages which are really
under 4G in machine physical addresses (only in virtual pseudo-physical
addresses).  To work around this, implement a vmalloc variant which
allocates each page with dma_alloc_coherent() to guarantee that each
page is suitable for the device in question.

(This is playing rather fast-and-loose with dma_alloc_coherent by
assuming that vfree will correctly release the pages.  That happens
to be OK for Xen, but will quite likely break if its drawing memory
from another type of IOMMU... So I don't think this alone is the correct
solution.)

Signed-off-by: Jeremy Fitzhardinge <jeremy.fitzhardinge@citrix.com>