Disable HPET in Xen pv dom0
Xen is using HPET to wakeup cpu from deep c-states, so the HPET usage in dom0
must be fobidden. In 2.6.18-xen, the HPET was by default disabled in config
file. For upstream kernel built for pv dom0, it is not practical to disable dom0
HPET usage in that way because the same image may be also using as bare metal
kernerl. So do it in Xen specific setup code.

Signed-off-by: Wei Gang <gang.wei@intel.com>
Signed-off-by: Jeremy Fitzhardinge <jeremy.fitzhardinge@citrix.com>