Add hugepage support to balloon driver
This patch adds hugepage support to the balloon driver.  It is activated
by specifying "balloon_hugepages" on the kernel command line.  Once activated,
the balloon driver will work entirely in hugepage sized chunks.

If, when returning pages, it finds a hugepage that is not contiguous
at the machine level, it will return each underlying page separately.
When this page is later repopulated it will be contiguous.

Signed-off-by: Dave McCracken <dave.mccracken@oracle.com>

--------

 balloon.c |  171 +++++++++++++++++++++++++++++++++++++++++++++-----------------
 1 file changed, 125 insertions(+), 46 deletions(-)
Signed-off-by: Jeremy Fitzhardinge <jeremy.fitzhardinge@citrix.com>