xen/balloon: update 'page' for each page
Since the latest ballooning patch "Add hugepage support to balloon driver"
0e898d5e9b893324435e0e77b67fe8d2f0523dbe i experience dom0 console
freezes. Additional to that there is a stacktrace about a kernel-bug in
drivers/xen/balloon.c in line 346, invalide opcode 0000 (i can send the
full stacktrace if someone is interested).

The cause for that seems to be that in the decrease_reservation-funtion
the PageHighMem - check is not executed on the assocciated page for the
current pfn. Its always the last page that got allocated in the for-loop
before, when all the pages are allocated.

With the attached patch that crashes dont occur anymore. But im not a
kernel-developer so im not sure that this is the correct fix.

Signed-off-by: Ronny.Hegewald@online.de <Ronny.Hegewald@online.de>
Signed-off-by: Jeremy Fitzhardinge <jeremy.fitzhardinge@citrix.com>
Acked-by: Dave McCracken <dcm@mccr.org>