xen: support MAXSMP configurations
The MAXSMP config option requires CPUMASK_OFFSTACK, which in turn
requires we init the memory for the maps while we bringing up the cpus.
MAXSMP also increases NR_CPUS to 4096. This increase in size exposed an
issue in the argument construction for mulitcalls from
xen_flush_tlb_others. The args should only need space for the actual
number of cpus, which with xen is currently only up to 32.

Signed-off-by: Andrew Jones <drjones@redhat.com>
Signed-off-by: Jeremy Fitzhardinge <jeremy.fitzhardinge@citrix.com>