x86: In PV guest mode, initialize MSI/MSI-X via the pci-front driver.
When we run Linux as PV guest with PCI pass-through we have to make
sure to call the pci-front driver to initialize the PCI card on our
behalf.

We call the xen_setup_msi_irqs which calls the pci-front driver
functions (if they are registered). When de-registering, we call
xen_teardown_msi_dev.