ttm: When TTM_PAGE_FLAG_DMA32 allocate pages under 4GB under Xen.
Under Xen, the flags for alloc_page (say, __GFP_DMA32) are ignored.
This is b/c the MMU notion of PFNs is an illusion - there is another
mapping of PFN to the real PFN, called MFNs.

This patch calls dma_alloc_coherent, which ends its way in
xen_swiotlb_alloc_coherent, which makes a call to alloc_page
but then swizzles the MFN under the PFN to be under the DMA witdh
(32-bit).

Signed-off-by: Konrad Rzeszutek Wilk <konrad.wilk@oracle.com>