xen-pcifront: Enforce scanning of device functions on initial execution.
'pci_scan_slot' abondons scanning of functions above 0 if a device with
function has not been detected. We need to be able to scan functions
above 0 in case the user has passed in devices without the function 0
for the slot/bus. To that end we are reusing the code that existed in
the rescan code path and make usage of it in the initial execution
path.

Signed-off-by: Konrad Rzeszutek Wilk <konrad.wilk@oracle.com>