agp: Use Xen back-door to get bus address for legacy code.
The AGP drivers for newer chipsets (ICH9 and higher) use
the PCI DMA and get the proper bus address.

For older "legacy" code (say ICH5), this has not been done.
To make those chipsets work properly, we need to program the
bus address of the pages in GATT with the correct address.

When running under Xen, the old trick of PFN << PAGE_SIZE == phys
does not work as the PFN is not neccessary equal to the "real" hardware
PFN (called 'MFN'). As such we either need to use the PCI DMA which
does the proper virt to bus address translation or use a backdoor
to do the translation.

This hack^H^H^patch chooses the second and MUST be revisted in the future.

Signed-off-by: Konrad Rzeszutek Wilk <konrad.wilk@oracle.com>