agp: Program the GART with the real physical address under Xen.
Before this patch, under Xen, we would program the GART with a
physical address that we thought would point to the GATT was actually
pointing to a random location. The end result was that the GATT
would fetch physical address for its virtual address translation from
a random memory location and end up ..well, making the video card
not work.

This uses a Xen back-door to figure out the real physical address.

We MUST revisit this to see about using the PCI DMA.

Signed-off-by: Konrad Rzeszutek Wilk <konrad.wilk@oracle.com>