xen/gntdev: use spinlocks rather than rwsem for locking
The mmu notifier mechanism calls its callbacks with an rcu lock,
which disables preemption.  This means we cannot use any blocking
synchronization for locking.

Convert all the rwsemas to plain spinlocks.  This requires that
the memory allocation and copying to/from userspace be split
from the actual datastructure updates since they can't be done
under spinlock.

Signed-off-by: Stefano Stabellini <Stefano.Stabellini@eu.citrix.com>
Signed-off-by: Jeremy Fitzhardinge <jeremy.fitzhardinge@citrix.com>