ttm: Change VMA flags if they != to the TTM flags.
During PCI suspend the graphic cards purge their VRAM and
move their graphic objects to the TT. They also unmap all of the
objects, meaning that when an user application is unfrozen it will
re-fault and call here.

What this means is that the VMA for the graphic object might
have been set for VRAM TTM but now it is with the TT (normal RAM)
meaning that the vma->vm_flags could be inappropiate (say, VM_IO
on TT - which is no good).

Virtual Computer (http://www.virtualcomputer.com) provided this
patch to the community.

Signed-off-by: Konrad Rzeszutek Wilk <konrad.wilk@oracle.com>