x86/spinlocks: replace pvspinlocks with pv ticketlocks
Rather than outright replacing the entire spinlock implementation in
order to paravirtualize it, keep the ticket lock implementation but add
a couple of slow-path pvops hooks.

Ticket locks have a number of nice properties, but they also have some
surprising behaviours.  They enforce a strict FIFO ordering on cpus
trying to take a lock; however, if the hypervisor scheduler does not
wake the (v)cpus in the correct order, the system can waste a huge amount
of time spinning until the next cpu can take the lock.

To address this, we add two hooks:
 - One in __ticket_spin_lock which is called after the cpu has been
   spinning on the lock for a certain number of iterations but has failed
   to take the lock.  The lock_spinning pvop is expected to block the
   cpu until it is woken, and
 - One in __ticket_spin_unlock, which on releasing a contended lock
   (there are more cpus with tail tickets), it looks to see if the next
   cpu is blocked and wakes it if so.

When compiled with CONFIG_PARAVIRT_SPINLOCKS disabled, a set of stub
functions causes all the extra code to go away.

Signed-off-by: Jeremy Fitzhardinge <jeremy.fitzhardinge@citrix.com>