x86/spinlocks: convert spin loop to C
The inner loop of __ticket_spin_lock isn't doing anything very special,
so reimplement it in C.  For the 8 bit ticket lock variant, we use a
register union to get direct access to the lower and upper bytes in the
tickets, but unfortunately gcc won't generate a direct comparison between
the two halves of the register.  We can, however, generate proper reads
into the lower half.

Signed-off-by: Jeremy Fitzhardinge <jeremy.fitzhardinge@citrix.com>