pvhvm: fix PV on HVM on xen/next-2.6.32
This patch makes PV on HVM work on xen/next-2.6.32: gsi and msi
remapping is not supported when running as HVM guests yet.

I am working on irq remapping for HVM guests and I have a prototype
working already but I think this simple fix is worth applying now anyway
to fix HVM guests with this kernel as soon as possible.

Signed-off-by: Stefano Stabellini <stefano.stabellini@eu.citrix.com>
Signed-off-by: Jeremy Fitzhardinge <jeremy.fitzhardinge@citrix.com>