Fix one race condition for netfront smartpoll logic
Assume one case like, netfront poll could not get any data,
and it clears the shared flag to indicate it is not polling.
However at this moment (netfront has cleared the flag but
still in hrtimer callback), netback triggers one interrupt
to netfront and tries to start the timer, the operation will
be failed since the timer is still alive.

Add this logic that, if the new hrtimer starts failed,
netfront should clear the shared flag to indicate that it is
not polling.

Signed-off-by: Dongxiao Xu <dongxiao.xu@intel.com>
Signed-off-by: Jeremy Fitzhardinge <jeremy.fitzhardinge@citrix.com>